agno-building-ai-agents/agno-teams/pyproject.toml
[project]
name = "agno-teams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.2.1",
    "ddgs>=9.6.1",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
